testing: add support for benchsplit flag
Split each benchmark into s intervals, which are run sequentially,
each for at least benchtime/s.

To avoid excessive ramp up time, initial operation counts after the
first interval are estimated based on cumulative time and operation
counts over already executed intervals.

**Example output**
With benchsplit=5:
BenchmarkSleepForOneMillisecond-8 (#0)	     200	   1222501 ns/op
BenchmarkSleepForOneMillisecond-8 (#1)	     200	   1236915 ns/op
BenchmarkSleepForOneMillisecond-8 (#2)	     200	   1233699 ns/op
BenchmarkSleepForOneMillisecond-8 (#3)	     200	   1237680 ns/op
BenchmarkSleepForOneMillisecond-8 (#4)	     200	   1238873 ns/op

With benchsplit=1 (default old behavior; no change in output):
BenchmarkSleepForOneMillisecond-8   	    1000	   1238305 ns/op

Fixes #19128

+++ src/testing/benchmark_test.go
@@ -13,7 +13,7 @@ import (
 )
 
 var roundDownTests = []struct {
-	v, expected int
+	v, expected int64
 }{
 	{1, 1},
 	{9, 1},
@@ -36,7 +36,7 @@ func TestRoundDown10(t *testing.T) {
 }
 
 var roundUpTests = []struct {
-	v, expected int
+	v, expected int64
 }{
 	{0, 1},
 	{1, 1},
